pyproject.toml
[tool.poetry]
name = "mealdeal-backend"
version = "0.1.0"
description = ""
authors = ["Kalle Sundell <sundellkalle@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "3.0.3"
pyodbc = "5.1.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
